pyproject.toml
[project]
name = "rag-in-a-box"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-search-documents>=11.6.0",
    "beautifulsoup4>=4.14.3",
    "gradio>=6.2.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
